init.qcom.post_boot: UNFUCK Trinket
sched_{up|down}migrate can only be valid as an array on 
Tri-cluster(Silver+Gold+Prime), which we are not. I have seen this on 
Laurel's DT for years now, so this is really a long time coming, someone 
was most likely copying blindly from sm8150 back then and thought 
this would improve a damn.

Another one that was here too was the background cpuset was set to CPU 0 to 5, 
which will assign 2 cores of Gold cluster(CPU 4 and 5) in our case, 
since our Silver has only 4 core(0-3), that someone must have been copying 
from another DT with a 6+2 CPU, breaking the whole purpose of the background groups.

Let's minimally unfuck all of these fuckeries as I will be using this ROM soon, 
and I also don't want other people to have any kind of bad impression on 
first boot, without No.912.

Signed-off-by: Fieah Lim <kweifat@gmail.com>